+++ tests/testthat/test-examples.R
@@ -1,6 +1,5 @@
 context("examples")
 
-source("utils.R")
 
 # some helpers
 run_example <- function(example) {